feat("feat(parser): first commit, add ability to performe cyclic dma with a ring buffer port"): When entering the dma interrupt, regardless of whatever value is read from the status register of the DMA IP core, enter the callback function
When entering the dma interrupt, regardless of whatever value is read from the status register of the DMA IP core, enter the callback function
------------------------------------------------------------
The loopback test is performed in the same thread, the incremental data is written to the TX Buffer, the received data is read from the RX Buffer and stored in an array. After the transfer is complete, the data in the array will be written to the file.
Even if the memcopy operation cannot handle the data in real time, since the increment operation and memcopy are performed at the same position, the data will still not be discontinuous. The first data may be duplicated because the memcopy is not done immediately after the DMA transfer is started.
Through the hardware flow control of the AXIS bus on the PL side, a throughput of 36MB/s can pass the test.

BREAKING CHANGE: Copyright (C) 2022 Guanzhou Elec. Inc.